+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 8/9/26
 
 - Special pages now warn about missing alt texts, same as posts
+- Special pages now warn about title+name both set, or title with no content
 
 ### 6/9/26